requirements.txt
# 必需依赖
lark-oapi>=1.0.0
flask>=2.0.0
qrcode[pil]>=7.0
schedule>=1.0.0
cryptography>=3.0
pillow>=8.0

# 可选：Dashboard Resources（AWS EC2/RDS 自动发现 + CloudWatch 指标）
# boto3>=1.20.0

# 可选：AWS CLI（Kiro Skill 中执行 AWS 命令）
# awscli>=2.0.0
